able to read the file again

+++ cpp/src/arrow/type.h
@@ -679,7 +679,7 @@ class ARROW_EXPORT BaseBinaryType : public DataType {
 
 constexpr int64_t kBinaryMemoryLimit = std::numeric_limits<int32_t>::max() - 1;
 
-constexpr int64_t kLargeBinaryMemoryLimit = std::numeric_limits<int32_t>::max() - 1; // 2^35
+constexpr int64_t kLargeBinaryMemoryLimit = std::numeric_limits<int64_t>::max() - 1; // 2^35
 
 /// \addtogroup binary-datatypes
 ///
